code-analyzer-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1